+++ lib/features/game/cubit/game_state.dart
@@ -12,6 +12,6 @@ class GameState with _$GameState {
     @Default(false) bool gameHasBegun,
     @Default(false) bool playerWin,
     @Default(0) gameStartTimeInMilliSecSinceEpoch,
-    @Default(50) int boxSize,
+    @Default(50) double boxSize,
   }) = _GameState;
 }
